requirements.txt
google-cloud-storage
google-cloud-firestore
google-genai
functions-framework
flask